+++ test-generated/entry/mainAny.test.js
@@ -2,7 +2,7 @@ import assert from 'assert'
 import * as mainAny from '../../src/entry/mainAny'
 import * as factoriesAny from '../../src/factoriesAny'
 import { createSnapshotFromFactories, validateBundle, validateTypeOf } from '../../src/utils/snapshot'
-const { create, all, add, matrix, isObject, isMatrix, pi, speedOfLight, sqrt, evaluate, chain, reviver, Complex } = mainAny
+const { create, all, add, matrix, isObject, isMatrix, pi, speedOfLight, sqrt, evaluate, chain, reviver, Complex, addDependencies } = mainAny
 
 const {
   expectedInstanceStructure,
@@ -24,7 +24,7 @@ describe('mainAny', function () {
     // snapshot testing
     const newMathInstance = create(all)
 
-    // don't output all warnings "math.foo.bar is move to math.bar, ..."
+    // don't output all deprecation warnings "math.foo.bar is move to math.bar, ..."
     const originalWarn = console.warn
     console.warn = (...args) => {
       if (args.join(' ').indexOf('is moved to') === -1) {
@@ -35,6 +35,35 @@ describe('mainAny', function () {
     validateBundle(expectedInstanceStructure, newMathInstance)
 
     console.warn = originalWarn
+  })
+
+  it('new instance should import all factory functions via import', function () {
+    // snapshot testing
+    const newMathInstance = create()
+
+    newMathInstance.import(all)
+
+    // don't output all deprecation warnings "math.foo.bar is move to math.bar, ..."
+    const originalWarn = console.warn
+    console.warn = (...args) => {
+      if (args.join(' ').indexOf('is moved to') === -1) {
+        originalWarn.apply(console, args)
+      }
+    }
+
+    validateBundle(expectedInstanceStructure, newMathInstance)
+
+    console.warn = originalWarn
+  })
+
+  it.only('new instance should import some factory functions via import', function () {
+    const newMathInstance = create()
+
+    newMathInstance.import({
+      addDependencies
+    }, { silent: true })
+
+    assert.strictEqual(newMathInstance.add(2, 3), 5)
   })
 
   it('evaluate should contain all functions from mathWithTransform', function () {
@@ -137,7 +166,6 @@ describe('mainAny', function () {
     assert.deepStrictEqual(obj, c)
   })
 
-  // TODO: test export of create and core
   // TODO: test export of errors
   // TODO: test export of classes
 })